requirements.txt
starlette==0.46.1
uvicorn==0.34.0
fastapi==0.115.11
kagglehub[pandas-datasets]
SQLAlchemy==2.0.39
marshmallow-sqlalchemy==1.4.1
PyJWT==2.10.1
bcrypt==4.3.0
pytest
pytest-cov